# backend-go/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Iran Gold Predictor API
  description: >
    Public Go API for the Iran gold price analysis platform. All timestamps are
    UTC ISO-8601. Iranian monetary values are IRT (toman). Predictions are
    uncertain estimates, not financial advice. Errors use the envelope
    {"error":{"code","message","details"}}. Every response carries X-Request-ID.
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /api/v1/readiness:
    get:
      tags: [system]
      summary: Readiness check (db + redis)
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ready }
                  db: { type: boolean }
                  redis: { type: boolean }
        "503":
          $ref: "#/components/responses/Error"
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register a user
      description: >
        First registered user becomes admin. Later registrations require an
        admin bearer token unless ALLOW_OPEN_REGISTRATION=true. Password must
        be at least 10 characters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
      responses:
        "201":
          description: Created (same envelope as login)
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expires_at: { type: string, format: date-time }
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in (rate limited to 10/min per IP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expires_at: { type: string, format: date-time }
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Current user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/prices/current:
    get:
      tags: [market]
      summary: Latest price per symbol with 24h change and staleness
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current prices keyed by canonical symbol
          content:
            application/json:
              schema:
                type: object
                properties:
                  as_of: { type: string, format: date-time }
                  prices:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/CurrentPrice"
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/prices/history:
    get:
      tags: [market]
      summary: Price history with pagination and optional bucketing
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: symbol, in: query, required: true, schema: { $ref: "#/components/schemas/Symbol" } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: interval, in: query, schema: { type: string, enum: [raw, hourly, daily], default: raw } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 500 } }
      responses:
        "200":
          description: Paged price items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        observed_at: { type: string, format: date-time }
                        value: { type: number }
                        source: { type: string }
                  page: { type: integer }
                  page_size: { type: integer }
                  total: { type: integer }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/market/summary:
    get:
      tags: [market]
      summary: Market snapshot (18k, xau, usd_irt, premium, providers, latest signal)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  as_of: { type: string, format: date-time }
                  current_18k: { $ref: "#/components/schemas/SummaryPrice" }
                  xau_usd: { $ref: "#/components/schemas/SummaryPrice" }
                  usd_irt: { $ref: "#/components/schemas/SummaryPrice" }
                  theoretical_18k: { type: number, nullable: true }
                  premium_pct: { type: number, nullable: true }
                  premium_avg_30d: { type: number, nullable: true }
                  last_update: { type: string, format: date-time, nullable: true }
                  providers:
                    type: array
                    items: { $ref: "#/components/schemas/ProviderHealth" }
                  signal:
                    allOf: [{ $ref: "#/components/schemas/Signal" }]
                    nullable: true
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/market/premium:
    get:
      tags: [market]
      summary: Daily observed vs theoretical 18k price and premium series
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: days, in: query, schema: { type: integer, minimum: 1, maximum: 3650, default: 90 } }
      responses:
        "200":
          description: Premium series
          content:
            application/json:
              schema:
                type: object
                properties:
                  days: { type: integer }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string, example: "2026-07-20" }
                        observed_18k: { type: number, nullable: true }
                        theoretical_18k: { type: number, nullable: true }
                        premium_pct: { type: number, nullable: true }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/market/indicators:
    get:
      tags: [market]
      summary: Technical indicators computed in Go over daily IR_GOLD_18K series
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: days, in: query, schema: { type: integer, minimum: 1, maximum: 3650, default: 90 } }
      responses:
        "200":
          description: Indicators (latest values + daily series)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Indicators" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/market/provider-gap:
    get:
      tags: [market]
      summary: Dispersion between providers quoting the same symbol
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: symbol, in: query, schema: { type: string, default: IR_GOLD_18K } }
        - { name: window_minutes, in: query, schema: { type: integer, minimum: 5, maximum: 1440, default: 120 } }
        - { name: history_days, in: query, schema: { type: integer, minimum: 0, maximum: 180, default: 30 } }
      responses:
        "200":
          description: >-
            Latest good quote per provider inside the window, the current gap
            ((max-min)/median %), and a daily gap-history series.
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/issues:
    get:
      tags: [issues]
      summary: Aggregated warnings/errors from all services (Issues tab)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 100 } }
        - { name: level, in: query, schema: { type: string, enum: [warning, error] } }
        - { name: service, in: query, schema: { type: string, enum: [api, prediction, frontend] } }
        - { name: since_hours, in: query, schema: { type: integer, minimum: 1, maximum: 720, default: 72 } }
      responses:
        "200":
          description: Issue rows, newest first
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
    post:
      tags: [issues]
      summary: Report a frontend error into the shared issue log
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                level: { type: string, enum: [warning, error], default: error }
                source: { type: string }
                message: { type: string }
                details: { type: object }
      responses:
        "202": { description: Accepted (queued for asynchronous insert) }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/issues/report:
    get:
      tags: [issues]
      summary: Markdown debug digest (recent issues + provider health + training runs)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Self-contained Markdown report for pasting into a debugging conversation
          content:
            text/markdown:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/predictions:
    get:
      tags: [predictions]
      summary: Latest prediction per horizon
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Latest prediction per horizon
          content:
            application/json:
              schema:
                type: object
                properties:
                  as_of: { type: string, format: date-time }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Prediction" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/predictions/custom:
    get:
      tags: [predictions]
      summary: On-demand forecast for an arbitrary 1-90 day horizon (not persisted)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: days, in: query, required: true, schema: { type: integer, minimum: 1, maximum: 90 } }
      responses:
        "200":
          description: >-
            Live-computed forecast: point/interval/direction/confidence, the
            winning fast-candidate model, and a cost-aware decision_lean
            (buy/hold/sell) with explanation. Ephemeral.
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/predictions/{horizon}:
    get:
      tags: [predictions]
      summary: Prediction history for a horizon (includes actual_value)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: horizon, in: path, required: true, schema: { $ref: "#/components/schemas/Horizon" } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 50 } }
      responses:
        "200":
          description: Prediction history
          content:
            application/json:
              schema:
                type: object
                properties:
                  horizon: { $ref: "#/components/schemas/Horizon" }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Prediction" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/signals/current:
    get:
      tags: [signals]
      summary: Latest trading signal for one symbol
      description: >
        Readings are published for seven signal-eligible symbols. ?symbol=
        defaults to IR_GOLD_18K, and the response shape is unchanged from
        before multi-asset support: no `symbol` field is added, because the
        caller already knows which symbol it asked for and the existing
        consumers of this payload must keep working byte for byte. An unknown
        symbol, or one that is not signal-eligible (DXY is an index, US10Y a
        yield, IR_GOLD_FUND_FLOW a ratio), is refused with 400 and the reason.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: symbol, in: query, schema: { $ref: "#/components/schemas/SignalSymbol" } }
      responses:
        "200":
          description: Latest signal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Signal" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/signals/history:
    get:
      tags: [signals]
      summary: Signal history for one symbol
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: symbol, in: query, schema: { $ref: "#/components/schemas/SignalSymbol" } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 50 } }
      responses:
        "200":
          description: Signal list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Signal" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/signals/overview:
    get:
      tags: [signals]
      summary: Latest reading per signal-eligible symbol
      description: >
        One item per eligible symbol, carrying the display metadata joined from
        the instrument registry, whether a trained model stood behind the
        reading, the factors that could NOT be scored and why, and whether the
        round-trip cost hurdle was observed or assumed.


        An item is returned only while its reading is CURRENT: nothing older
        than 24 hours is served, and anything past the engine's own six-hour
        review window carries `reading_stale` with its age in
        `reading_age_hours`. `as_of` is the newest reading in the response, not
        a timestamp common to all of them, so `oldest_reading_at` publishes the
        other end of the range.


        `unavailable` accounts for everything this board does not publish, each
        entry saying which KIND of absence it is in `category`: an eligible
        symbol with no current reading, a symbol missing from the instrument
        registry, a registered symbol this service collects but deliberately
        does not score (DXY, US10Y, BRENT_OIL, IR_GOLD_FUND_FLOW,
        IR_GOLD_FUND_KAHRABA), and the asset classes it never collects at all
        (cars, housing, Tehran-listed equities, local Iranian silver). The
        category matters because the reasons are not interchangeable — the
        first four of those symbols are collected daily, so "not collected"
        would be the opposite of the truth about them.


        Nothing here is computed by this service; it is a projection of what
        prediction-python wrote.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Multi-asset overview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignalOverview" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/models:
    get:
      tags: [models]
      summary: Model versions (active + recent)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Model versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ModelVersion" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/models/performance:
    get:
      tags: [models]
      summary: Per-horizon active model metrics vs baseline plus live accuracy
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Performance summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  horizons:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        model: { $ref: "#/components/schemas/ModelVersion" }
                        metrics: { type: object }
                        baseline_metrics: { type: object }
                        live:
                          type: object
                          nullable: true
                          properties:
                            n_matured: { type: integer }
                            mape_pct: { type: number, nullable: true }
                            directional_accuracy: { type: number, nullable: true }
                            interval_coverage: { type: number, nullable: true }
                  last_training_run:
                    type: object
                    nullable: true
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/portfolio:
    get:
      tags: [portfolio]
      summary: Holdings and computed valuation
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Portfolio with computed summary (flat fields)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PortfolioSummary"
                  - type: object
                    properties:
                      holdings:
                        type: array
                        items: { $ref: "#/components/schemas/Transaction" }
                      current_price: { type: number }
                      price_available: { type: boolean }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/portfolio/transactions:
    post:
      tags: [portfolio]
      summary: Add a transaction
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer, format: int64 }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/portfolio/transactions/{id}:
    put:
      tags: [portfolio]
      summary: Update a transaction
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionInput" }
      responses:
        "200": { description: Updated }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [portfolio]
      summary: Delete a transaction
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200": { description: Deleted }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/portfolio/import:
    post:
      tags: [portfolio]
      summary: Import transactions from CSV (multipart, max 1MB)
      description: >
        Required columns: tx_type,grams,karat,price_per_gram,currency,fees,tx_date,notes.
        Rows with formula-injection cells (leading = + - @) are rejected. The
        import is all-or-nothing.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported: { type: integer }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
  /api/v1/portfolio/export:
    get:
      tags: [portfolio]
      summary: Export transactions as CSV (formula-injection sanitized)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    Symbol:
      type: string
      enum: [IR_GOLD_18K, XAUUSD, XAGUSD, USD_IRT, IR_COIN_EMAMI, BRENT_OIL, DXY, US10Y]
    Horizon:
      type: string
      enum: [1h, 4h, eod, 1d, 3d, 7d, 30d]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { type: string, enum: [user, admin] }
    CurrentPrice:
      type: object
      properties:
        value: { type: number }
        currency: { type: string }
        unit: { type: string }
        source: { type: string }
        observed_at: { type: string, format: date-time }
        stale:
          type: boolean
          description: >
            Market-hours aware: while the market is open, true when older than
            STALE_MINUTES; while closed, last-session data is not stale.
        market_state: { $ref: "#/components/schemas/MarketState" }
        change_24h_pct: { type: number, nullable: true }
...